+++ changelog.md
@@ -1,5 +1,8 @@
 # 0.3.3
 
+### New Features
+* Get state of noble with `.isNobleReady()`
+
 ### Bug Fixes
 * Fix bug where scan stop would not stop scan which led to many problems. Especially when it comes